backend/requirements.txt
# # blinker==1.9.0
# # click==8.2.1
# # flask==3.1.1
# # flask-cors==6.0.0
# # itsdangerous==2.2.0
# # Jinja2==3.1.6
# # MarkupSafe==3.0.2
# # PyJWT==2.10.1
# # Werkzeug==3.1.3

# # Flask==3.1.1
# # flask-cors==6.0.0
# # Flask-SQLAlchemy==3.1.1
# # psycopg2-binary==2.9.9
# # PyJWT==2.10.1


# Flask==3.1.1
# flask-cors==6.0.0
# Flask-SQLAlchemy==3.1.1
# psycopg2-binary==2.9.9
# PyJWT==2.10.1
# sentence-transformers==2.2.2
# scikit-learn==1.3.2


Flask==3.1.1
flask-cors==6.0.0
pymongo==4.5.0
dnspython==2.3.0
PyJWT==2.10.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
